+++ packages/rainbowkit/src/components/ConnectOptions/ConnectDetails.tsx
@@ -74,7 +74,7 @@ export function GetDetail({
   const wallets = useWalletConnectors();
   const shownWallets = wallets.splice(0, 5);
 
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   return (
     <Box
@@ -214,7 +214,7 @@ export function ConnectDetail({
   const getDesktopDeepLink = wallet.desktop?.getUri;
   const safari = isSafari();
 
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   const hasExtension = !!wallet.extensionDownloadUrl;
   const hasQrCodeAndExtension = downloadUrls?.qrCode && hasExtension;
@@ -625,7 +625,7 @@ export function DownloadOptionsDetail({
     mobileDownloadUrl,
   } = wallet;
 
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   useEffect(() => {
     // Preload icons used on next screen
@@ -732,7 +732,7 @@ export function DownloadDetail({
 }) {
   const { downloadUrls, qrCode } = wallet;
 
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   useEffect(() => {
     // Preload icons used on next screen
@@ -813,7 +813,7 @@ export function InstructionMobileDetail({
   connectWallet: (wallet: WalletConnector) => void;
   wallet: WalletConnector;
 }) {
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   return (
     <Box
@@ -900,7 +900,7 @@ export function InstructionExtensionDetail({
 }: {
   wallet: WalletConnector;
 }) {
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   return (
     <Box
@@ -989,7 +989,7 @@ export function InstructionDesktopDetail({
   connectWallet: (wallet: WalletConnector) => void;
   wallet: WalletConnector;
 }) {
-  const i18n = useContext(I18nContext);
+  const { i18n } = useContext(I18nContext);
 
   return (
     <Box